Login
Users are able to login and a user object is received, converted to
JSON and saved in Shared Properties

+++ FestiFind/FestiFind/EventLocation.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 public class EventLocation : Location {
-    override init(longitude: Int, latitude: Int) {
+    override init(longitude: Double, latitude: Double) {
         super.init(longitude: longitude, latitude: latitude)
     }
 }
